Add js,goback and modify main java crs navbar code

+++ src/main/java/com/StudManageSys/api/StudentController.java
@@ -36,11 +36,32 @@ public class StudentController {
 	                             @RequestParam("branch") String branch,
 	                             @RequestParam("email") String email,
 	                             @RequestParam("dob") String dob,
-	                             @RequestParam("gender") String gender) {
+	                             @RequestParam("gender") String gender, Model model) {
 	        Student student = new Student(kodID, name, branch, email, dob, gender);
-	        ss.addStudent(student);
+	        boolean idExists=ss.checkId(kodID);
+	        boolean emailExists=ss.checkEmail(email);
+	        if(idExists==false) {
+	        	if(emailExists==false) {
+	        		ss.addStudent(student);
+	        System.out.println("Successfull Added !");
+	        String result = ss.addSuccessMassage();
+	        model.addAttribute("result",result);
 	        return "register";
+        	}else {
+        		System.out.println("ID exists");
+        		String result = ss.addExists();
+	        	model.addAttribute("result",result);
+        		return "register";
+        	}
+	        }else {
+	        	System.out.println("email exists");
+	        	String result = ss.addExists();
+	        	model.addAttribute("result",result);
+	        	return "register";
+	        }
+	        
 	    }
+	    
 
 	    @GetMapping("/viewInfo")
 	    public String viewInfo() {
@@ -50,9 +71,14 @@ public class StudentController {
 
 	    @GetMapping("/view")
 	    public String getStudent(@RequestParam("kodID") String kodID, Model model) {
+	    	try {
 	        Student student = ss.getStudent(kodID);
+	        System.out.println("successfully Get");
 	        model.addAttribute("student", student);
 	        return "showInfo";
+	    	}catch(Exception e){
+	    		return "goback";
+	    	}
 	    }
 
 		@GetMapping("/showAllStudents")
@@ -67,6 +93,7 @@ public class StudentController {
 	        model.addAttribute("students", students);
 	        return "showAllStudents";
 	    	}catch(Exception e) {
+	    		System.out.println("Not available database");
 	    		return "index";
 	    	}
 	    }
@@ -82,10 +109,22 @@ public class StudentController {
 	                                @RequestParam("branch") String branch,
 	                                @RequestParam("email") String email,
 		                            @RequestParam("dob") String dob,
-		                            @RequestParam("gender") String gender) {
+		                            @RequestParam("gender") String gender, Model model) {
 	        Student student = new Student(kodID, name, branch, email, dob, gender);
+	        boolean idExists=ss.checkId(kodID);
+	        if(idExists==true) {
 	        ss.updateStudent(student);
-	        return "updateInfo";
+	        String result = ss.updateSuccessMassage();
+	        model.addAttribute("result1",result);
+	        return "updateInfo";	
+	        }else {
+	        	String result = ss.updateExists();
+		        model.addAttribute("result1",result);
+	        	return "updateInfo";
+	        }
+	        
+	        
+	        
 	    }	
 	    @GetMapping("/remove")
 	    public String removeStudent() {
@@ -93,9 +132,19 @@ public class StudentController {
 	    }
 
 	    @PostMapping("/deleteStudent")
-	    public String deleteStudent(@RequestParam("kodID") String kodID) {
-	        ss.deleteStudent(kodID);
-	        return "remove";
+	    public String deleteStudent(@RequestParam("kodID") String kodID, Model model) {
+	    	boolean idExists=ss.checkId(kodID);
+	    	if(idExists==true) {
+	    	ss.deleteStudent(kodID);
+	    	String result = ss.removeSuccessMassage();
+	    	model.addAttribute("result2",result);
+	        return "remove";	
+	    	}else {
+	    		String result = ss.removeExists();
+		    	model.addAttribute("result2",result);
+	    		return "remove";
+	    	}
+	        
 	        
 	    }
